Reorganize files to create Python package
