+++ hazardapp/templates/base.html
