+++ hazardapp/templates/home.html
