+++ hazardapp/templates/input.html
